third_party/WebKit/LayoutTests/external/wpt/web-animations/animation-model/animation-types/discrete-animation-expected.txt

This is a testharness.js-based test.
PASS Test animating discrete values
PASS Test discrete animation is used when interpolation fails
PASS Test discrete animation is used only for pairs of values that cannot be interpolated
FAIL Test the 50% switch point for discrete animation is based on the effect easing assert_equals: Animation produces 'to' value at 96% of the iteration time expected "oblique" but got "italic"
FAIL Test the 50% switch point for discrete animation is based on the keyframe easing assert_equals: Animation produces 'to' value at 96% of the iteration time expected "oblique" but got "italic"
Harness: the test ran to completion.
